Change configuration type for FST max_size

+++ src/config/config.rs
@@ -137,7 +137,7 @@ pub struct ConfigStoreFSTGraph {
     pub consolidate_after: u64,
 
     #[serde(default = "defaults::store_fst_graph_max_size")]
-    pub max_size: usize,
+    pub max_size: u64,
 
     #[serde(default = "defaults::store_fst_graph_max_words")]
     pub max_words: usize,
